requirements/development.txt
# Develoment dependencies
# -----------------------

black

flake8-builtins>=2.2
flake8-eradicate>=1.5
flake8-isort>=6.1
flake8-qgis>=1

isort>=5.13
pre-commit>=3,<4
